// src/frontend/src/app/api/fn/exams/api-exams-exam-id-patch-json.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { ExamDto } from '../../models/exam-dto';

export interface ApiExamsExamIdPatch$Json$Params {
  examId: number;
      body?: {
'ExamName'?: string;
'SqlFile'?: Blob;
}
}

export function apiExamsExamIdPatch$Json(http: HttpClient, rootUrl: string, params: ApiExamsExamIdPatch$Json$Params, context?: HttpContext): Observable<StrictHttpResponse<ExamDto>> {
  const rb = new RequestBuilder(rootUrl, apiExamsExamIdPatch$Json.PATH, 'patch');
  if (params) {
    rb.path('examId', params.examId, {});
    rb.body(params.body, 'multipart/form-data');
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'text/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<ExamDto>;
    })
  );
}

apiExamsExamIdPatch$Json.PATH = '/api/exams/{examId}';